Autobuild added, data hook added too.

+++ build.py
@@ -4,41 +4,57 @@ import urllib.request
 import zipfile
 
 
-# Download zips using "url"
-print("Downloading static file zip.")
+def build_assets():
 
-url = (
-    "https://github.com/alphagov/govuk-frontend/"
-    "releases/download/v4.0.0/release-v4.0.0.zip"
-)
+    if os.path.exists("app/static"):
 
-# There is a known problem on Mac where one must manually
-# run the script "Install Certificates.command" found
-# in the python application folder for this to work.
-urllib.request.urlretrieve(url, "./govuk_frontend.zip")
+        print(
+            "Assets already built."
+            "If you require a rebuild manually run build.build_assets"
+        )
 
-print("Deleting old app/static")
+        return True
 
-# Attempts to delete the old files, states if
-# one doesnt exist.
-try:
-    shutil.rmtree("app/static")
-except FileNotFoundError:
-    print("No old app/static to remove.")
+    # Download zips using "url"
+    print("Downloading static file zip.")
 
-print("Unzipping file to app/static...")
+    url = (
+        "https://github.com/alphagov/govuk-frontend/"
+        "releases/download/v4.0.0/release-v4.0.0.zip"
+    )
 
-# Extracts the previously downloaded zip to /app/static
-with zipfile.ZipFile("./govuk_frontend.zip", "r") as zip_ref:
-    zip_ref.extractall("./app/static")
+    # There is a known problem on Mac where one must manually
+    # run the script "Install Certificates.command" found
+    # in the python application folder for this to work.
+    urllib.request.urlretrieve(url, "./govuk_frontend.zip")  # nosec
 
-print("Moving files from app/static/assets to app/static")
+    print("Deleting old app/static")
 
-for file_to_move in os.listdir("./app/static/assets"):
-    shutil.move("./app/static/assets/" + file_to_move, "app/static")
+    # Attempts to delete the old files, states if
+    # one doesnt exist.
+    try:
+        shutil.rmtree("app/static")
+    except FileNotFoundError:
+        print("No old app/static to remove.")
 
-print("Deleting app/static/assets")
+    print("Unzipping file to app/static...")
 
-# Deletes temp. files.
-shutil.rmtree("./app/static/assets")
-os.remove("./govuk_frontend.zip")
+    # Extracts the previously downloaded zip to /app/static
+    with zipfile.ZipFile("./govuk_frontend.zip", "r") as zip_ref:
+        zip_ref.extractall("./app/static")
+
+    print("Moving files from app/static/assets to app/static")
+
+    for file_to_move in os.listdir("./app/static/assets"):
+        shutil.move("./app/static/assets/" + file_to_move, "app/static")
+
+    print("Deleting app/static/assets")
+
+    # Deletes temp. files.
+    shutil.rmtree("./app/static/assets")
+    os.remove("./govuk_frontend.zip")
+
+
+if __name__ == "__main__":
+
+    build_assets()
